feature[#10]: 만든 앨범수 조회 구현

+++ src/main/kotlin/com/vibetrip/vibetripserver/member/presentation/dto/response/MemberResponse.kt
@@ -4,6 +4,6 @@ data class MemberResponse (
     val name: String,
     val email: String,
     val profilImage : String,
-    val albumCount: Int,
-    val albumLogCount: Int,
+    val albumCount: Long,
+    val albumLogCount: Long,
 )
